requirements.txt
python-telegram-bot
requests
beautifulsoup4
plotly
python-dotenv
kaleido
lxml